requirements.txt
autopep8
channels-redis
django-autoslug
django-channels
django-contrib-comments
djoser
docutils
drf-yasg
mysqlclient
nplusone
pillow
pip-chill
pyopenssl
